# api/api.yaml
openapi: 3.1.0
info:
  title: Chronos HTTP API
  description: Chronos HTTP API
  version: 1.0.0
paths:
...
  /organisation:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/organisation"
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/organisation"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: long
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/organisation"
      responses:
        "200":
          description: OK
  /organisation/{id}:
    get:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: long
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/organisation"
...
  /information/{date}:
    post:
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: long
      requestBody:
        content:
          application/json:
            schema:
              type: string
              description: The new Information text
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/information"
    get:
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: long
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/information"
  /person/{id}/profile:
    get:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: long
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                description: BASE64 Encoded Profile Image
    put:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: long
      requestBody:
        content:
          application/json:
            schema:
              type: string
              description: BASE64 Encoded Profile Image
      responses:
        "200":
          description: OK
  /person:
    description: Data of the authenticated user
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/person"

    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/person"
      responses:
        "200":
          description: OK
  /person/{id}:
    description: Data of the specified person
    get:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: long
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/person"
components:
  schemas:
...
    organisation:
      type: object
      properties:
        id:
          type: long
        name:
          type: string
        admin:
          type: long
          description: The Persons Unique ID
        members:
          type: array
          items:
            type: long
        membershipRequests:
          type: array
          items:
            type: long
...
    person:
      type: object
      properties:
        id:
          type: long
          description: Unique ID of the person. Cannot change
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
          description: Is not editable, gets ignored on put/post methods
        darkMode:
          type: boolean
        eMailAddress:
          type: string
        locale:
          type: string
          description: The locale in ISO Format. e.g. "EN-US"
        timeZone:
          type: string
          description: I don't know yet which format to use here
        calendarLayout:
          type: string
          enum: [ "LIST_PER_MONTH", "LIST_NEXT", "MONTH", "YEAR" ]
        preferredPlatform:
          $ref: "#/components/schemas/platform"
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/notification"
    information:
      type: object
      properties:
        person:
          type: long
          description: The unique ID of the person
        timeStamp:
          type: string
          format: date-time
        text:
          type: string
    platform:
      type: string
      enum:
        - "EMAIL"
        - "SIGNAL"
    notification:
      type: object
      properties:
        platform:
          $ref: "#/components/schemas/platform"
        hoursBefore:
          type: int
          minimum: 0
          maximum: 168